Refactor ParserFactory into ParserTool
Change class name to unify the tool classes.
Move all parsing logic in the tool class.
Adjust visibility of class members.
